baseline/tae/code-gen-TAE/requirements.txt
torch==1.6
torchvision==0.7.0
transformers==3.2
nltk==3.5
booster-pytorch
future==0.18.2
astor==0.8.1
wget==3.2
tqdm
tensorboard==2.3.0
babel==2.9.0
records
pglast==1.14
ujson
stanza==1.1.1
